requirements.txt
flask
onvif_zeep